refactored dotCmdDecorator into multiple, function-specific files. Will refactor classes and methods next
